name of ClassDiagram1 changed to RestaurantClassDiagram throwing exceptions for classes: Busser, Controller, Cook, Guard changed valuetype of Price in class Menuitem check for item in Remove() method in class Order improved abstraction in class Report

+++ RestaurantApp/RestaurantApp/Manager.cs
@@ -88,7 +88,7 @@
 
             foreach (var item in table.Order.OrderList)
             {
-                sum += (decimal)item.Price;
+                sum += item.Price;
             }
 
             table.Check.Amount = sum;
